0_data_preparation/requirements.txt
google-cloud-bigquery
pandas
requests
beautifulsoup4
lxml
html5lib